cleanup
https://coveord.atlassian.net/browse/KIT-4966
