+++ packages/atomic/src/components/commerce/atomic-product-numeric-field-value/atomic-product-numeric-field-value.ts
@@ -1,7 +1,7 @@
 import {html, LitElement} from 'lit';
 import {customElement, property, state} from 'lit/decorators.js';
 import {when} from 'lit/directives/when.js';
-import {createProductContextController} from '@/src/components/commerce/product-template-component-utils/product-template-controllers.js';
+import {createProductContextController} from '@/src/components/commerce/product-template-component-utils/context/product-context-controller.js';
 import {bindingGuard} from '@/src/decorators/binding-guard.js';
 import {bindings} from '@/src/decorators/bindings.js';
 import {errorGuard} from '@/src/decorators/error-guard.js';
