+++ packages/atomic/src/components/commerce/product-template-component-utils/context/stencil-product-template-decorators.tsx
@@ -1,5 +1,5 @@
 import {Product} from '@coveo/headless/commerce';
-import {itemContext} from '../../common/item-list/stencil-item-decorators';
+import {itemContext} from '@/src/components/common/item-list/stencil-item-decorators';
 
 /**
  * Retrieves `Product` on a rendered `atomic-product`.
@@ -11,7 +11,7 @@ import {itemContext} from '../../common/item-list/stencil-item-decorators';
  * @param element - The element that the event is dispatched to, which must be the child of a rendered "atomic-product".
  * @returns A promise that resolves on initialization of the parent "atomic-product" element, or rejects when there is no parent "atomic-product" element.
  *
- * @deprecated should only be used for Stencil components. For Lit components, use `fetchProductContext` from \@/src/components/commerce/product-template-component-utils/product-template-controllers.ts
+ * @deprecated should only be used for Stencil components. For Lit components, use `fetchProductContext` from \@/src/components/commerce/product-template-component-utils/context/fetch-product-context.ts
  */
 export function productContext<T extends Product>(element: Element) {
   return itemContext<T>(element, 'atomic-product');
