+++ packages/atomic/src/components/insight/atomic-insight-result-quickview-action/atomic-insight-result-quickview-action.tsx
@@ -18,7 +18,7 @@ import {
 } from '../../../utils/stencil-accessibility-utils';
 import {IconButton} from '../../common/stencil-iconButton';
 import {Bindings} from '../../search/atomic-search-interface/atomic-search-interface';
-import {ResultContext} from '@/src/components/search/result-template-component-utils/stencil-result-template-decorators';
+import {ResultContext} from '@/src/components/search/result-template-component-utils/context/stencil-result-template-decorators';
 
 /**
  * @internal
