+++ packages/atomic/src/components/search/result-template-components/atomic-result-date/atomic-result-date.tsx
@@ -9,7 +9,7 @@ import {
   InitializeBindings,
 } from '../../../../utils/initialization-utils';
 import {Bindings} from '../../atomic-search-interface/atomic-search-interface';
-import {ResultContext} from '@/src/components/search/result-template-component-utils/stencil-result-template-decorators';
+import {ResultContext} from '@/src/components/search/result-template-component-utils/context/stencil-result-template-decorators';
 
 dayjs.extend(calendar);
 dayjs.extend(updateLocale);
