+++ packages/atomic/src/components/search/result-template-components/atomic-result-localized-text/atomic-result-localized-text.ts
@@ -1,7 +1,7 @@
 import {isNullOrUndefined} from '@coveo/bueno';
 import {type Result, ResultTemplatesHelpers} from '@coveo/headless';
 import {Component, Prop, State} from '@stencil/core';
-import {ResultContext} from '@/src/components/search/result-template-component-utils/stencil-result-template-decorators';
+import {ResultContext} from '@/src/components/search/result-template-component-utils/context/stencil-result-template-decorators';
 import {
   type InitializableComponent,
   InitializeBindings,
