+++ packages/atomic/src/components/search/result-template-components/atomic-result-timespan/atomic-result-timespan.tsx
@@ -6,7 +6,7 @@ import {DurationUnitType} from 'dayjs/plugin/duration';
 import duration from 'dayjs/plugin/duration';
 import {InitializeBindings} from '../../../../utils/initialization-utils';
 import {Bindings} from '../../atomic-search-interface/atomic-search-interface';
-import {ResultContext} from '@/src/components/search/result-template-component-utils/stencil-result-template-decorators';
+import {ResultContext} from '@/src/components/search/result-template-component-utils/context/stencil-result-template-decorators';
 
 dayjs.extend(duration);
