+++ packages/atomic/src/index.ts
@@ -2,8 +2,8 @@ export type {i18n} from 'i18next';
 export {Components, JSX} from './components';
 export {CommerceBindings} from './components/commerce/atomic-commerce-interface/atomic-commerce-interface';
 export {SelectChildProductEventArgs} from './components/commerce/atomic-product-children/select-child-product-event';
-export {fetchProductContext} from './components/commerce/product-template-component-utils/product-template-controllers';
-export {productContext} from './components/commerce/product-template-component-utils/stencil-product-template-decorators';
+export {fetchProductContext} from './components/commerce/product-template-component-utils/context/fetch-product-context';
+export {productContext} from './components/commerce/product-template-component-utils/context/stencil-product-template-decorators';
 export {PopoverChildFacet} from './components/common/facets/popover/popover-type';
 export {RedirectionPayload} from './components/common/search-box/redirection-payload';
 export {dispatchSearchBoxSuggestionsEvent} from './components/common/suggestions/suggestions-events';
@@ -19,7 +19,7 @@ export {
 } from './components/common/suggestions/suggestions-utils';
 export {RecsBindings} from './components/recommendations/atomic-recs-interface/atomic-recs-interface';
 export {Bindings} from './components/search/atomic-search-interface/atomic-search-interface';
-export {resultContext} from './components/search/result-template-component-utils/stencil-result-template-decorators';
+export {resultContext} from './components/search/result-template-component-utils/context/stencil-result-template-decorators';
 export {
   initializeBindings,
   MissingInterfaceParentError,
